requirements.txt
aiofiles==23.2.1
fastapi==0.111.0
Jinja2==3.1.4
openai==1.26.0
python-dotenv==1.0.1
python-multipart==0.0.9
uvicorn==0.29.0
websockets==12.0